requirements.txt
fastapi
uvicorn
pynacl
requests